Fix memorySize calculation for quantiles

+++ presto-main/src/main/java/com/facebook/presto/operator/aggregation/state/DigestAndPercentileArrayStateFactory.java
@@ -91,7 +91,7 @@ public class DigestAndPercentileArrayStateFactory
         }
 
         @Override
-        public void addMemoryUsage(int value)
+        public void addMemoryUsage(long value)
         {
             size += value;
         }
@@ -135,7 +135,7 @@ public class DigestAndPercentileArrayStateFactory
         }
 
         @Override
-        public void addMemoryUsage(int value)
+        public void addMemoryUsage(long value)
         {
             // noop
         }
