+++ presto-main/src/main/java/com/facebook/presto/operator/aggregation/state/DigestAndPercentileState.java
@@ -29,5 +29,5 @@ public interface DigestAndPercentileState
 
     void setPercentile(double percentile);
 
-    void addMemoryUsage(int value);
+    void addMemoryUsage(long value);
 }
